refactor(targets): Simplify target construction
Remove the unused abstract target interface and make runner target construction explicitly create LettaCodeTarget.
