+++ letta_evals/targets/__init__.py
@@ -1,4 +1,4 @@
-from letta_evals.targets.base import AbstractAgentTarget
+from letta_evals.targets.errors import TargetError
 from letta_evals.targets.letta_code_target import LettaCodeTarget
 
-__all__ = ["AbstractAgentTarget", "LettaCodeTarget"]
+__all__ = ["LettaCodeTarget", "TargetError"]
